+++ changelog.md
@@ -1,6 +1,6 @@
 # 0.4.0.0
 
-* Add `foldr`.
+* Add `foldr` catamorphism and `fromInfinite` / `toInfinite` conversions.
 * Remove instances `Foldable` and `Traversable`, they are too dangerous to diverge.
 
 # 0.3.3.0